delete修正＆シフト登録追加

+++ index.js
@@ -361,6 +361,26 @@ app.get('/shiftdata', async (req, res) => {
     res.json('')
   }
 })
+app.post('/shiftdatabyid', async (req, res) => {
+  try {
+    const id = req.body[0]
+    const client = await pool.connect()
+    const result = await client.query(
+      'select * from shift_table where userid=$1',
+      [id],
+    )
+    const results = { results: result ? result.rows : null }
+    if (result.rowCount == 0) {
+      res.send([])
+    } else {
+      res.json(results.results)
+    }
+    client.release()
+  } catch (err) {
+    console.error(err)
+    res.send('Error ' + err)
+  }
+})
 app.post('/addshiftdata', async (req, res) => {
   if (req.session.access_token != null) {
     try {
@@ -388,6 +408,31 @@ app.post('/addshiftdata', async (req, res) => {
     res.json('')
   }
 })
+app.post('/addshiftdatabyid', async (req, res) => {
+  try {
+    const data = req.body
+    const id = data.id
+    const shifts = data.shift
+    const client = await pool.connect()
+    for (let shift of shifts) {
+      const result = await client.query(
+        'insert into shift_table (userid,date,detail) values($1,$2,$3)',
+        [id, shift.date, shift.text],
+      )
+      writeLog(
+        client,
+        id,
+        '管理者によるシフト追加',
+        shift.date + '[' + shift.text + ']',
+      )
+    }
+    client.release()
+    res.json('add')
+  } catch (err) {
+    console.error(err)
+    res.send('Error ' + err)
+  }
+})
 app.post('/deleteshiftdata', async (req, res) => {
   if (req.session.access_token != null) {
     try {
@@ -395,8 +440,8 @@ app.post('/deleteshiftdata', async (req, res) => {
       const client = await pool.connect()
       for (let data of deldata) {
         const result = await client.query(
-          'delete from shift_table where date=$1',
-          [data.date],
+          'delete from shift_table where date=$1 and userid=$2',
+          [data.date, req.session.userId],
         )
         writeLog(client, req.session.userId, 'シフト削除', data.date)
       }
@@ -408,6 +453,26 @@ app.post('/deleteshiftdata', async (req, res) => {
     }
   } else {
     res.json('')
+  }
+})
+app.post('/deleteshiftdatabyid', async (req, res) => {
+  try {
+    const data = req.body
+    const id = data.id
+    const shifts = data.shift
+    const client = await pool.connect()
+    for (let shift of shifts) {
+      const result = await client.query(
+        'delete from shift_table where date=$1 and userid=$2',
+        [shift.date, id],
+      )
+      writeLog(client, id, '管理者によるシフト削除', shift.date)
+    }
+    client.release()
+    res.json('delete')
+  } catch (err) {
+    console.error(err)
+    res.send('Error ' + err)
   }
 })
 app.post('/updateshiftdata', async (req, res) => {
@@ -435,6 +500,31 @@ app.post('/updateshiftdata', async (req, res) => {
     }
   } else {
     res.json('')
+  }
+})
+app.post('/updateshiftdatabyid', async (req, res) => {
+  try {
+    const data = req.body
+    const id = data.id
+    const shifts = data.shift
+    const client = await pool.connect()
+    for (let shift of shifts) {
+      const result = await client.query(
+        'update shift_table set detail=$1 where userid=$2 and date=$3',
+        [shift.text, id, shift.date],
+      )
+      writeLog(
+        client,
+        id,
+        '管理者によるシフト更新',
+        shift.date + '[' + shift.text + ']',
+      )
+    }
+    client.release()
+    res.json('update')
+  } catch (err) {
+    console.error(err)
+    res.send('Error ' + err)
   }
 })
 app.get('/allshiftdata', async (req, res) => {
